refactor(core): move DEFAULT_QUEUE_NAME and poolQueueName to queue-name.ts
