+++ packages/cli/src/scaling/pool-config.service.ts
@@ -4,7 +4,7 @@ import { jsonParse } from 'n8n-workflow';
 
 import { CacheService } from '@/services/cache/cache.service';
 
-import { poolQueueName } from './constants';
+import { poolQueueName } from './queue-name';
 import type { ExecutionCategory, PoolAssignment } from './scaling.types';
 
 const SETTINGS_KEY = 'workerPools.assignment';
