+++ packages/cli/src/scaling/queue-name.ts
@@ -1,10 +1,15 @@
 import type { InstanceSettings } from 'n8n-core';
 
-import { DEFAULT_QUEUE_NAME } from './constants';
+export const DEFAULT_QUEUE_NAME = 'jobs';
 
 type InstanceType = InstanceSettings['instanceType'];
 
 export function resolveQueueName(instanceType: InstanceType, poolName: string): string {
 	if (instanceType !== 'worker' || poolName === '') return DEFAULT_QUEUE_NAME;
+	return `${DEFAULT_QUEUE_NAME}-${poolName}`;
+}
+
+export function poolQueueName(poolName: string | undefined): string {
+	if (!poolName) return DEFAULT_QUEUE_NAME;
 	return `${DEFAULT_QUEUE_NAME}-${poolName}`;
 }
